refactor: restructure multimodal model forward functions and optimize code style
- Add multimodal_forwards module to centrally manage multimodal model forward logic
- Extract and optimize forward function implementations for Qwen2 VL and Qwen2.5 VL
- Improve sequence_parallel related code structure
