+++ src/llamafactory/data/processors/sequence_parallel.py
@@ -75,8 +75,6 @@ def sp_split(examples, model_args, tokenizer):
     for k, v in examples.items():
         chunks = list()
         for row in v:
-            #if k.endswith("attention_mask"):
-            #    chunks.extend([row] * model_args.sequence_parallel_size)
             if row is None:
                 chunks.extend([None] * model_args.sequence_parallel_size)
             elif k in ['images']:
